+++ backend/langgraph.json
@@ -1,0 +1,161 @@
+{
+  "graphs": {
+    "main": "app.server:app"
+  },
+  "dependencies": [
+    "aiohappyeyeballs",
+    "aiohttp",
+    "aiosignal",
+    "annotated-types",
+    "antlr4-python3-runtime",
+    "anyio",
+    "attrs",
+    "backoff",
+    "beautifulsoup4",
+    "cachetools",
+    "certifi",
+    "cffi",
+    "chardet",
+    "charset-normalizer",
+    "click",
+    "coloredlogs",
+    "contourpy",
+    "cryptography",
+    "cycler",
+    "dataclasses-json",
+    "deprecated",
+    "distro",
+    "effdet",
+    "emoji",
+    "eval-type-backport",
+    "faiss-cpu",
+    "fastapi",
+    "filelock",
+    "filetype",
+    "flatbuffers",
+    "fonttools",
+    "frozenlist",
+    "fsspec",
+    "google-api-core",
+    "google-auth",
+    "google-cloud-vision",
+    "googleapis-common-protos",
+    "grpcio",
+    "grpcio-status",
+    "h11",
+    "html5lib",
+    "httpcore",
+    "httpx",
+    "huggingface-hub",
+    "humanfriendly",
+    "idna",
+    "inflection",
+    "iopath",
+    "jinja2",
+    "jiter",
+    "joblib",
+    "jsonpatch",
+    "jsonpath-python",
+    "jsonpointer",
+    "jsonschema",
+    "jsonschema-specifications",
+    "kiwisolver",
+    "langchain",
+    "langchain-cohere",
+    "langchain-community",
+    "langchain-core",
+    "langchain-openai",
+    "langchain-text-splitters",
+    "langdetect",
+    "langsmith",
+    "layoutparser",
+    "lxml",
+    "markupsafe",
+    "marshmallow",
+    "matplotlib",
+    "mpmath",
+    "multidict",
+    "mypy-extensions",
+    "nest-asyncio",
+    "nltk",
+    "numpy",
+    "olefile",
+    "omegaconf",
+    "onnx",
+    "openai",
+    "openapi",
+    "opencv-python",
+    "orjson",
+    "packaging",
+    "pandas",
+    "pdf2image",
+    "pdfminer-six",
+    "pdfplumber",
+    "pi-heif",
+    "pikepdf",
+    "pillow",
+    "portalocker",
+    "propcache",
+    "proto-plus",
+    "protobuf",
+    "psutil",
+    "pyasn1",
+    "pyasn1-modules",
+    "pycocotools",
+    "pycparser",
+    "pyparsing",
+    "pypdf",
+    "pypdfium2",
+    "python-dateutil",
+    "python-dotenv",
+    "python-iso639",
+    "python-magic",
+    "python-multipart",
+    "python-oxmsg",
+    "pytz",
+    "pyyaml",
+    "rapidfuzz",
+    "referencing",
+    "regex",
+    "requests",
+    "requests-toolbelt",
+    "rpds-py",
+    "rsa",
+    "safetensors",
+    "setuptools",
+    "six",
+    "sniffio",
+    "soupsieve",
+    "gunicorn",
+    "sqlalchemy",
+    "starlette",
+    "sympy",
+    "tenacity",
+    "tiktoken",
+    "timm",
+    "tokenizers",
+    "transformers",
+    "typing-inspect",
+    "typing-extensions",
+    "tzdata",
+    "unstructured",
+    "unstructured-client",
+    "unstructured-inference",
+    "unstructured-pytesseract",
+    "urllib3",
+    "uvicorn",
+    "webencodings",
+    "wrapt",
+    "yarl",
+    "youtube-transcript-api",
+    "yt-dlp",
+    "scikit-learn",
+    "ariadne",
+    "boto3",
+    "botocore",
+    "moto",
+    "langchain-neo4j",
+    "langgraph",
+    "langgraph-cli[inmem]"
+  ]
+}
