+++ backend/pyproject.toml
@@ -164,7 +164,8 @@ dependencies = [
     "botocore (>=1.34.0)",
     "moto (>=4.2.0)",
     "langchain-neo4j (>=0.4.0,<0.5.0)",
-    "langgraph (>=0.5.0,<0.6.0)"
+    "langgraph (>=0.5.0,<0.6.0)",
+    "langgraph-cli[inmem] (>=0.3.3,<0.4.0)"
 ]
